backend/requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
python-dotenv==1.0.1
pydantic==2.12.0
requests==2.31.0
gunicorn==21.2.0
flask-cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
flask-marshmallow==1.2.1
marshmallow-sqlalchemy==1.4.2
passlib[bcrypt]==1.7.4
email-validator==2.1.1
joblib>=1.3.0
numpy>=1.26.0
pandas>=2.0.0
Pillow>=10.0.0
scikit-learn>=1.3.0
